move revalidate to lib
